# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UpdateUserBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"birthday": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"organizationId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": {}
			},
			"ProductCreateBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"external",
							"internal"
						]
					},
					"costPrice": {
						"type": "number",
						"format": "double"
					},
					"sellPrice": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"supplier": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"type",
					"costPrice",
					"sellPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductUpdateBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"external",
							"internal"
						]
					},
					"costPrice": {
						"type": "number",
						"format": "double"
					},
					"sellPrice": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"supplier": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"OrderCreateBody": {
				"properties": {
					"customerId": {
						"type": "string"
					},
					"productId": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string",
						"enum": [
							"USD",
							"VND"
						]
					}
				},
				"required": [
					"customerId",
					"productId",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExpenseCreateBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string",
						"enum": [
							"USD",
							"VND"
						]
					},
					"source": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"amount",
					"source"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CustomerCreateBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"staffId": {
						"type": "string"
					},
					"leadSourceId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustomerUpdateBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"staffId": {
						"type": "string"
					},
					"leadSourceId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ChangePasswordBody": {
				"properties": {
					"password": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"cfNewPassword": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SendMailBody": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "tc-ecommerce",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/user/{userId}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Lấy thông tin chi tiết một user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Cập nhật thông tin user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserBody"
							}
						}
					}
				}
			}
		},
...
		"/product": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xem danh sách tất cả sản phẩm",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "supplier",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Tạo sản phẩm mới (chỉ Admin)",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductCreateBody"
							}
						}
					}
				}
			}
		},
		"/product/{productId}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xem chi tiết sản phẩm",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Sửa sản phẩm (chỉ Admin)",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductUpdateBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xoá sản phẩm (chỉ Admin)",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xem danh sách đơn hàng\nStaff chỉ thấy đơn hàng của khách hàng do mình quản lý",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"pending",
								"completed",
								"cancelled"
							]
						}
					},
					{
						"in": "query",
						"name": "staffId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Tạo đơn hàng mới",
				"tags": [
					"Order"
				],
				"security": [
					{
						"jwt": [
							"ADMIN",
							"STAFF"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrderCreateBody"
							}
						}
					}
				}
			}
		},
		"/expense": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xem danh sách chi tiêu công ty",
				"tags": [
					"Expense"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Tạo chi tiêu mới",
				"tags": [
					"Expense"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExpenseCreateBody"
							}
						}
					}
				}
			}
		},
...
		"/customer": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xem danh sách khách hàng",
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "staffId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "leadSourceId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Tạo khách hàng mới",
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": [
							"ADMIN",
							"STAFF"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CustomerCreateBody"
							}
						}
					}
				}
			}
		},
		"/customer/{customerId}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xem thông tin một khách hàng",
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "customerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Cập nhật thông tin khách hàng",
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": [
							"ADMIN",
							"STAFF"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "customerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CustomerUpdateBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Xoá khách hàng (Chỉ Admin)",
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "customerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/auth/change-password": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Đổi mật khẩu",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordBody"
							}
						}
					}
				}
			}
		},
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Quên mật khẩu",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordBody"
							}
						}
					}
				}
			}
		},
		"/auth/send-mail": {
			"post": {
				"operationId": "SendMail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Gửi mail xác thực",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendMailBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}